Split CBORGenerator into Custom.Generator and Custom.Validator
