+++ src/Codec/CBOR/Cuddle/CDDL/Custom/Generator.hs
@@ -1,65 +1,33 @@
 {-# LANGUAGE TypeData #-}
 {-# LANGUAGE TypeFamilies #-}
 
-module Codec.CBOR.Cuddle.CDDL.CBORGenerator (
-  MonadCddl (..),
+module Codec.CBOR.Cuddle.CDDL.Custom.Generator (
   GenPhase,
-  ValidatorPhase,
-
-  -- * Custom generators
   CBORGen,
+  XXCTree (..),
+  HasGenerator (..),
   GenConfig (..),
   GenEnv (..),
-  HasGenerator (..),
-  WrappedTerm (..),
-  XXCTree (..),
-  runCBORGen,
   liftAntiGen,
+  runCBORGen,
   withAntiGen,
-  withTwiddle,
   withLocalGenBindings,
-
-  -- * Custom validators
-  Validator,
-  TermValidator,
-  CustomValidatorResult (..),
-  ValidateEnv (..),
-  HasValidator (..),
-  runValidator,
-  withLocalValidateBindings,
+  disableTwiddle,
+  enableTwiddle,
 ) where
 
 import Codec.CBOR.Cuddle.CDDL (GRef (..), Name (..))
 import Codec.CBOR.Cuddle.CDDL.CTree (CTree, CTreeRoot (..), XXCTree)
-import Codec.CBOR.Term (Term)
-import Control.Monad.Reader (MonadReader (..), ReaderT (..), asks, local, mapReaderT)
-import Data.Kind (Type)
+import Codec.CBOR.Cuddle.CDDL.Custom.Core (MonadCddl (..), WrappedTerm)
+import Control.Monad.Reader (MonadReader (..), ReaderT (..), asks, mapReaderT)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
-import Data.Text (Text)
-import Data.Text qualified as T
 import GHC.Generics (Generic)
-import Optics.Core (Lens')
+import Optics.Lens (Lens')
 import Test.AntiGen (AntiGen)
 import Test.QuickCheck.GenT (MonadGen (..))
 
-class MonadCddl m where
-  type Phase m :: Type
-
-  -- | Look up a top-level rule by name.
-  lookupCddl :: Name -> m (Maybe (CTree (Phase m)))
-
-  -- | Look up the rule bound to a generic parameter at the enclosing rule.
-  -- Returns 'Nothing' outside of a custom generator/validator that was
-  -- attached to a generic rule.
-  lookupGRef :: GRef -> m (Maybe (CTree (Phase m)))
-
 type data GenPhase
-type data ValidatorPhase
-
-data instance XXCTree ValidatorPhase
-  = VRuleRef Name
-  | VValidator (WrappedTerm -> Validator ()) (CTree ValidatorPhase)
 
 -- | User-supplied configuration for the generator monad. Set once at the
 -- top level when calling 'runCBORGen'.
@@ -113,6 +81,12 @@ withTwiddle :: Bool -> CBORGen a -> CBORGen a
 withTwiddle t =
   local (\env -> env {geConfig = (geConfig env) {gcTwiddle = t}})
 
+disableTwiddle :: CBORGen a -> CBORGen a
+disableTwiddle = withTwiddle False
+
+enableTwiddle :: CBORGen a -> CBORGen a
+enableTwiddle = withTwiddle True
+
 -- | Run an action with the given local generic bindings installed.
 -- Used to wrap custom generators attached to generic rules so that
 -- 'lookupGRef' resolves to the type bound at the enclosing rule.
@@ -125,58 +99,5 @@ data instance XXCTree GenPhase
   = GenRef Name
   | GenGenerator (CBORGen WrappedTerm) (CTree GenPhase)
 
-data WrappedTerm
-  = -- | Single term
-    S Term
-  | -- | Pair term
-    P Term Term
-  | -- | Group term
-    G [WrappedTerm]
-  deriving (Eq, Ord, Show)
-
 class HasGenerator a where
   generatorL :: Lens' a (Maybe (CBORGen WrappedTerm))
-
-class HasValidator a where
-  validatorL :: Lens' a (Maybe (WrappedTerm -> Validator ()))
-
-data CustomValidatorResult
-  = CustomValidatorSuccess
-  | CustomValidatorFailure Text
-  deriving (Generic, Show, Eq)
-
-data ValidateEnv = ValidateEnv
-  { veRoot :: CTreeRoot ValidatorPhase
-  , veLocal :: Map Name (CTree ValidatorPhase)
-  }
-
-newtype Validator a = Validator (ReaderT ValidateEnv (Either Text) a)
-  deriving (Functor, Applicative, Monad, MonadReader ValidateEnv)
-
-type TermValidator = WrappedTerm -> Validator ()
-
-instance MonadCddl Validator where
-  type Phase Validator = ValidatorPhase
-
-  lookupCddl n = do
-    CTreeRoot root <- asks veRoot
-    pure $ Map.lookup n root
-
-  lookupGRef (GRef t) = do
-    binds <- asks veLocal
-    pure $ Map.lookup (Name t) binds
-
-instance MonadFail Validator where
-  fail msg = Validator . ReaderT $ \_ -> Left $ T.pack msg
-
--- | Run an action with the given local generic bindings installed.
-withLocalValidateBindings ::
-  Map Name (CTree ValidatorPhase) -> Validator a -> Validator a
-withLocalValidateBindings binds =
-  local (\env -> env {veLocal = binds `Map.union` veLocal env})
-
-runValidator ::
-  Validator () -> CTreeRoot ValidatorPhase -> CustomValidatorResult
-runValidator (Validator (ReaderT f)) cddl =
-  either CustomValidatorFailure (const CustomValidatorSuccess) $
-    f ValidateEnv {veRoot = cddl, veLocal = Map.empty}
